CIPHERSCHOOLS_Project/requirements.txt
requests
colorama
PyYAML 